LPS-64091 implement only for WABs, provide a proxy where we can hang the modifiable parts of the ServletContext specifically for ServletContainerInitializer && ServletContextListener
