+++ modules/apps/foundation/portal-osgi-web/portal-osgi-web-portlet-tracker/src/main/java/com/liferay/portal/osgi/web/portlet/tracker/internal/PortletTracker.java
@@ -924,6 +924,8 @@ public class PortletTracker
 		ServletContextHelperRegistration servletContextHelperRegistration =
 			getServletContextHelperRegistration(bundle, serviceRegistrations);
 
+		servletContextHelperRegistration.initDefaults();
+
 		ServletContext servletContext =
 			servletContextHelperRegistration.getServletContext();
