+++ modules/apps/foundation/portal-osgi-web/portal-osgi-web-servlet-context-helper/src/main/java/com/liferay/portal/osgi/web/servlet/context/helper/ServletContextHelperRegistration.java
@@ -37,6 +37,8 @@ public interface ServletContextHelperRegistration {
 	public ServiceReference<ServletContextListener>
 		getServletContextListenerSeviceReference();
 
+	public void initDefaults();
+
 	public void setProperties(Map<String, String> contextParameters);
 
 }
