+++ modules/apps/foundation/portal-osgi-web/portal-osgi-web-servlet-context-helper/src/main/java/com/liferay/portal/osgi/web/servlet/context/helper/internal/ServletContextHelperRegistrationImpl.java
@@ -16,36 +16,25 @@ package com.liferay.portal.osgi.web.servlet.context.helper.internal;
 
 import com.liferay.portal.kernel.portlet.RestrictPortletServletRequest;
 import com.liferay.portal.kernel.servlet.PortletServlet;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.osgi.web.servlet.context.helper.ServletContextHelperRegistration;
 import com.liferay.portal.osgi.web.servlet.jsp.compiler.JspServlet;
 
 import java.io.IOException;
-import java.io.InputStream;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 
 import java.net.URL;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Dictionary;
 import java.util.Enumeration;
-import java.util.HashSet;
 import java.util.Hashtable;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Set;
 
 import javax.portlet.PortletRequest;
 
@@ -55,24 +44,19 @@ import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
 import javax.servlet.Servlet;
 import javax.servlet.ServletConfig;
-import javax.servlet.ServletContainerInitializer;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletContextListener;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
-import javax.servlet.annotation.HandlesTypes;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import org.apache.felix.utils.log.Logger;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
 import org.osgi.framework.ServiceRegistration;
-import org.osgi.framework.wiring.BundleWiring;
 import org.osgi.service.http.context.ServletContextHelper;
 import org.osgi.service.http.whiteboard.HttpWhiteboardConstants;
 
@@ -90,55 +74,46 @@ public class ServletContextHelperRegistrationImpl
 
 		String contextPath = getContextPath(bundle);
 
-		String servletContextName = getServletContextName(bundle, contextPath);
-
-		boolean wabShapedBundle = false;
+		_servletContextName = getServletContextName(bundle, contextPath);
 
 		URL url = bundle.getEntry("WEB-INF/");
 
 		if (url != null) {
-			wabShapedBundle = true;
+			_wabShapedBundle = true;
+		}
+		else {
+			_wabShapedBundle = false;
 		}
 
-		BundleContext bundleContext = bundle.getBundleContext();
-
-		_logger = new Logger(bundleContext);
+		_bundleContext = bundle.getBundleContext();
 
 		_customServletContextHelper = new CustomServletContextHelper(
-			bundle, wabShapedBundle);
+			bundle, _wabShapedBundle);
 
 		_servletContextHelperServiceRegistration = createServletContextHelper(
-			bundleContext, servletContextName, contextPath);
+			_bundleContext, _servletContextName, contextPath);
 
 		_servletContextListenerServiceRegistration =
-			createServletContextListener(bundleContext, servletContextName);
-
-		initServletContainerInitializers(
-			bundle, getServletContext(), wabShapedBundle);
-
-		_defaultServletServiceRegistration = createDefaultServlet(
-			bundleContext, servletContextName, wabShapedBundle);
-
-		_jspServletServiceRegistration = createJspServlet(
-			bundleContext, servletContextName);
-
-		_portletServletServiceRegistration = createPortletServlet(
-			bundleContext, servletContextName, wabShapedBundle);
-
-		_portletServletRequestFilterServiceRegistration =
-			createRestrictPortletServletRequestFilter(
-				bundleContext, servletContextName);
+			createServletContextListener(_bundleContext, _servletContextName);
 	}
 
 	@Override
 	public void close() {
-		_servletContextHelperServiceRegistration.unregister();
+		if (_servletContextHelperServiceRegistration != null) {
+			_servletContextHelperServiceRegistration.unregister();
+		}
 
-		_servletContextListenerServiceRegistration.unregister();
+		if (_servletContextListenerServiceRegistration != null) {
+			_servletContextListenerServiceRegistration.unregister();
+		}
 
-		_defaultServletServiceRegistration.unregister();
+		if (_defaultServletServiceRegistration != null) {
+			_defaultServletServiceRegistration.unregister();
+		}
 
-		_jspServletServiceRegistration.unregister();
+		if (_jspServletServiceRegistration != null) {
+			_jspServletServiceRegistration.unregister();
+		}
 
 		if (_portletServletServiceRegistration != null) {
 			_portletServletServiceRegistration.unregister();
@@ -146,6 +121,30 @@ public class ServletContextHelperRegistrationImpl
 
 		if (_portletServletRequestFilterServiceRegistration != null) {
 			_portletServletRequestFilterServiceRegistration.unregister();
+		}
+	}
+
+	@Override
+	public void initDefaults() {
+		if (_defaultServletServiceRegistration == null) {
+			_defaultServletServiceRegistration = createDefaultServlet(
+				_bundleContext, _servletContextName, _wabShapedBundle);
+		}
+
+		if (_jspServletServiceRegistration == null) {
+			_jspServletServiceRegistration = createJspServlet(
+				_bundleContext, _servletContextName);
+		}
+
+		if (_portletServletServiceRegistration == null) {
+			_portletServletServiceRegistration = createPortletServlet(
+				_bundleContext, _servletContextName, _wabShapedBundle);
+		}
+
+		if (_portletServletRequestFilterServiceRegistration == null) {
+			_portletServletRequestFilterServiceRegistration =
+				createRestrictPortletServletRequestFilter(
+					_bundleContext, _servletContextName);
 		}
 	}
 
@@ -194,121 +193,6 @@ public class ServletContextHelperRegistrationImpl
 		}
 
 		_servletContextHelperServiceRegistration.setProperties(properties);
-	}
-
-	protected void collectAnnotatedClasses(
-		String classResource, Bundle bundle, Class<?>[] handledTypesArray,
-		Set<Class<?>> annotatedClasses) {
-
-		String className = classResource.replaceAll("\\.class$", "");
-
-		className = className.replaceAll("/", ".");
-
-		Class<?> annotatedClass = null;
-
-		try {
-			annotatedClass = bundle.loadClass(className);
-		}
-		catch (Throwable t) {
-			_logger.log(Logger.LOG_DEBUG, t.getMessage());
-
-			return;
-		}
-
-		// Class extends/implements
-
-		for (Class<?> handledType : handledTypesArray) {
-			if (handledType.isAssignableFrom(annotatedClass)) {
-				annotatedClasses.add(annotatedClass);
-
-				return;
-			}
-		}
-
-		// Class annotation
-
-		Annotation[] classAnnotations = new Annotation[0];
-
-		try {
-			classAnnotations = annotatedClass.getAnnotations();
-		}
-		catch (Throwable t) {
-			_logger.log(Logger.LOG_DEBUG, t.getMessage());
-		}
-
-		for (Annotation classAnnotation : classAnnotations) {
-			if (ArrayUtil.contains(
-					handledTypesArray, classAnnotation.annotationType())) {
-
-				annotatedClasses.add(annotatedClass);
-
-				return;
-			}
-		}
-
-		// Method annotation
-
-		Method[] classMethods = new Method[0];
-
-		try {
-			classMethods = annotatedClass.getDeclaredMethods();
-		}
-		catch (Throwable t) {
-			_logger.log(Logger.LOG_DEBUG, t.getMessage());
-		}
-
-		for (Method method : classMethods) {
-			Annotation[] methodAnnotations = new Annotation[0];
-
-			try {
-				methodAnnotations = method.getDeclaredAnnotations();
-			}
-			catch (Throwable t) {
-				_logger.log(Logger.LOG_DEBUG, t.getMessage());
-			}
-
-			for (Annotation methodAnnotation : methodAnnotations) {
-				if (ArrayUtil.contains(
-						handledTypesArray, methodAnnotation.annotationType())) {
-
-					annotatedClasses.add(annotatedClass);
-
-					return;
-				}
-			}
-		}
-
-		// Field annotation
-
-		Field[] declaredFields = new Field[0];
-
-		try {
-			declaredFields = annotatedClass.getDeclaredFields();
-		}
-		catch (Throwable t) {
-			_logger.log(Logger.LOG_DEBUG, t.getMessage());
-		}
-
-		for (Field field : declaredFields) {
-			Annotation[] fieldAnnotations = new Annotation[0];
-
-			try {
-				fieldAnnotations = field.getDeclaredAnnotations();
-			}
-			catch (Throwable t) {
-				_logger.log(Logger.LOG_DEBUG, t.getMessage());
-			}
-
-			for (Annotation fieldAnnotation : fieldAnnotations) {
-				if (ArrayUtil.contains(
-						handledTypesArray, fieldAnnotation.annotationType())) {
-
-					annotatedClasses.add(annotatedClass);
-
-					return;
-				}
-			}
-		}
 	}
 
 	protected String createContextSelectFilterString(
@@ -502,147 +386,28 @@ public class ServletContextHelperRegistrationImpl
 		return contextPath.replaceAll("[^a-zA-Z0-9\\-]", "");
 	}
 
-	protected void initServletContainerInitializers(
-		Bundle bundle, ServletContext servletContext, boolean wabShapedBundle) {
-
-		if (!wabShapedBundle) {
-			return;
-		}
-
-		BundleWiring bundleWiring = bundle.adapt(BundleWiring.class);
-
-		Collection<String> initializerResources = bundleWiring.listResources(
-			"META-INF/services", "javax.servlet.ServletContainerInitializer",
-			BundleWiring.LISTRESOURCES_RECURSE);
-
-		if (initializerResources == null) {
-			return;
-		}
-
-		for (String initializerResource : initializerResources) {
-			URL url = bundle.getResource(initializerResource);
-
-			if (url == null) {
-				continue;
-			}
-
-			try (InputStream inputStream = url.openStream()) {
-				String fqcn = StringUtil.read(inputStream);
-
-				processServletContainerInitializerClass(
-					fqcn, bundle, bundleWiring, servletContext);
-			}
-			catch (IOException ioe) {
-				_logger.log(Logger.LOG_ERROR, ioe.getMessage(), ioe);
-			}
-		}
-	}
-
-	protected void processServletContainerInitializerClass(
-		String fqcn, Bundle bundle, BundleWiring bundleWiring,
-		ServletContext servletContext) {
-
-		Class<? extends ServletContainerInitializer> initializerClass = null;
-
-		try {
-			Class<?> clazz = bundle.loadClass(fqcn);
-
-			if (!ServletContainerInitializer.class.isAssignableFrom(clazz)) {
-				return;
-			}
-
-			initializerClass = clazz.asSubclass(
-				ServletContainerInitializer.class);
-		}
-		catch (Exception e) {
-			_logger.log(Logger.LOG_ERROR, e.getMessage(), e);
-
-			return;
-		}
-
-		HandlesTypes handledTypesAnnotation = initializerClass.getAnnotation(
-			HandlesTypes.class);
-
-		if (handledTypesAnnotation == null) {
-			handledTypesAnnotation = _NULL_HANDLES_TYPES;
-		}
-
-		Class<?>[] handledTypesArray = handledTypesAnnotation.value();
-
-		if (handledTypesArray == null) {
-			handledTypesArray = new Class[0];
-		}
-
-		Collection<String> classResources = bundleWiring.listResources(
-			"/", "*.class", BundleWiring.LISTRESOURCES_RECURSE);
-
-		if (classResources == null) {
-			classResources = new ArrayList<>(0);
-		}
-
-		Set<Class<?>> annotatedClasses = new HashSet<>();
-
-		for (String classResource : classResources) {
-			URL urlClassResource = bundle.getResource(classResource);
-
-			if (urlClassResource == null) {
-				continue;
-			}
-
-			collectAnnotatedClasses(
-				classResource, bundle, handledTypesArray, annotatedClasses);
-		}
-
-		if (annotatedClasses.isEmpty()) {
-			annotatedClasses = null;
-		}
-
-		try {
-			ServletContainerInitializer servletContainerInitializer =
-				initializerClass.newInstance();
-
-			servletContainerInitializer.onStartup(
-				annotatedClasses, servletContext);
-		}
-		catch (Throwable t) {
-			_logger.log(Logger.LOG_ERROR, t.getMessage(), t);
-		}
-	}
-
 	private static final String _JSP_SERVLET_INIT_PARAM_PREFIX =
 		"jsp.servlet.init.param.";
-
-	private static final HandlesTypes _NULL_HANDLES_TYPES = new HandlesTypes() {
-
-		@Override
-		public Class<? extends Annotation> annotationType() {
-			return null;
-		}
-
-		@Override
-		public Class<?>[] value() {
-			return new Class[0];
-		}
-
-	};
 
 	private static final String _SERVLET_INIT_PARAM_PREFIX =
 		HttpWhiteboardConstants.HTTP_WHITEBOARD_SERVLET_INIT_PARAM_PREFIX;
 
-	private final CustomServletContextHelper _customServletContextHelper;
-	private final ServiceRegistration<?> _defaultServletServiceRegistration;
-	private final ServiceRegistration<Servlet> _jspServletServiceRegistration;
-	private final Logger _logger;
-	private final ServiceRegistration<?>
+	private final BundleContext _bundleContext;
+	private CustomServletContextHelper _customServletContextHelper;
+	private ServiceRegistration<?> _defaultServletServiceRegistration;
+	private ServiceRegistration<Servlet> _jspServletServiceRegistration;
+	private ServiceRegistration<?>
 		_portletServletRequestFilterServiceRegistration;
-	private final ServiceRegistration<Servlet>
+	private ServiceRegistration<Servlet>
 		_portletServletServiceRegistration;
 	private final Map<String, Object> _properties;
 	private final Props _props;
-	private final ServiceRegistration<ServletContextHelper>
+	private ServiceRegistration<ServletContextHelper>
 		_servletContextHelperServiceRegistration;
 	private final ServiceRegistration<ServletContextListener>
 		_servletContextListenerServiceRegistration;
+	private final String _servletContextName;
+	private final boolean _wabShapedBundle;
 
 	private static class JspServletWrapper extends HttpServlet {
